# hhspec/specs/architecture/api-contracts.yml
...
openapi: "3.0.3"
info:
  title: "WellAlly Health Web API"
  version: "1.0.0"
  description: "个人医疗健康数据管理 Web 界面后端 API"
servers:
  - url: "http://localhost:3000/api"
    description: "本地开发服务器"
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      operationId: register
      summary: 注册新账户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, display_name, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                display_name:
                  type: string
                  maxLength: 100
                password:
                  type: string
                  minLength: 6
      responses:
        201:
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        409:
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        200:
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        401:
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: 获取当前用户信息
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: 当前用户信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ==================== Reports ====================
...
  /consultations/specialist:
    post:
      operationId: createSpecialistConsult
      summary: 创建单专科咨询
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [specialty, chief_complaint]
              properties:
                specialty: { type: string, maxLength: 200 }
                chief_complaint: { type: string, maxLength: 5000 }
                related_report_ids: { type: array, items: { type: integer } }
      responses:
        201:
          description: 咨询已提交
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/mdt:
    post:
      operationId: createMDTConsult
      summary: 创建 MDT 多学科会诊
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [specialties, chief_complaint]
              properties:
                specialties:
                  type: array
                  items: { type: string }
                  minItems: 2
                chief_complaint: { type: string, maxLength: 5000 }
                related_report_ids: { type: array, items: { type: integer } }
      responses:
        201:
          description: 会诊已提交
  /consultations:
    get:
      operationId: listConsultations
      summary: 查询咨询历史
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: 咨询列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"

  # ==================== Trends ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        display_name: { type: string }
        role: { type: string }
        height_cm: { type: number, nullable: true }
        weight_kg: { type: number, nullable: true }
        birth_date: { type: string, format: date, nullable: true }
        gender: { type: string, nullable: true }
        blood_type: { type: string, nullable: true }
        api_key_configured: { type: boolean }
...
    Consultation:
      type: object
      properties:
        id: { type: integer }
        consultation_type: { type: string }
        specialty: { type: string, nullable: true }
        chief_complaint: { type: string }
        ai_response: { type: string, nullable: true }
        mdt_specialties: { type: array, items: { type: string }, nullable: true }
        mdt_summary: { type: string, nullable: true }
        status: { type: string }
        created_at: { type: string, format: date-time }
...
  responses:
...
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...